remove dataframe.py
