+++ src/snowflake/snowpark/modin/pandas/api/extensions/__init__.py
@@ -19,9 +19,12 @@
 # existing code originally distributed by the Modin project, under the Apache License,
 # Version 2.0.
 
-from modin.pandas.api.extensions import register_series_accessor
+from modin.pandas.api.extensions import (
+    register_dataframe_accessor,
+    register_series_accessor,
+)
 
-from .extensions import register_dataframe_accessor, register_pd_accessor
+from .extensions import register_pd_accessor
 
 __all__ = [
     "register_dataframe_accessor",
