+++ src/snowflake/snowpark/modin/pandas/indexing.py
@@ -45,6 +45,7 @@ import numpy as np
 import pandas
 from modin.pandas import Series
 from modin.pandas.base import BasePandasDataset
+from modin.pandas.dataframe import DataFrame
 from pandas._libs.tslibs import Resolution, parsing
 from pandas._typing import AnyArrayLike, Scalar
 from pandas.api.types import is_bool, is_list_like
@@ -60,7 +61,6 @@ from pandas.core.indexing import IndexingError
 
 import snowflake.snowpark.modin.pandas as pd
 import snowflake.snowpark.modin.pandas.utils as frontend_utils
-from snowflake.snowpark.modin.pandas.dataframe import DataFrame
 from snowflake.snowpark.modin.pandas.utils import is_scalar
 from snowflake.snowpark.modin.plugin._internal.indexing_utils import (
     MULTIPLE_ELLIPSIS_INDEXING_ERROR_MESSAGE,
