+++ src/snowflake/snowpark/modin/plugin/_internal/telemetry.py
@@ -9,7 +9,6 @@ from contextlib import nullcontext
 from enum import Enum, unique
 from typing import Any, Callable, Optional, TypeVar, Union, cast
 
-import modin.pandas
 from typing_extensions import ParamSpec
 
 import snowflake.snowpark.session
@@ -543,8 +542,6 @@ class TelemetryMeta(type):
     def __new__(
         cls, name: str, bases: tuple, attrs: dict[str, Any]
     ) -> Union[
-        "modin.pandas.Series",
-        "snowflake.snowpark.modin.pandas.dataframe.DataFrame",
         "snowflake.snowpark.modin.pandas.groupby.DataFrameGroupBy",
         "snowflake.snowpark.modin.pandas.resample.Resampler",
         "snowflake.snowpark.modin.pandas.window.Window",
@@ -558,7 +555,6 @@ class TelemetryMeta(type):
         with ``snowpark_pandas_telemetry_api_usage`` telemetry decorator.
         Method arguments returned by _get_kwargs_telemetry are collected otherwise set telemetry_args=list().
         TelemetryMeta is only set as the metaclass of:
-         snowflake.snowpark.modin.pandas.dataframe.DataFrame,
          snowflake.snowpark.modin.pandas.groupby.DataFrameGroupBy,
          snowflake.snowpark.modin.pandas.resample.Resampler,
          snowflake.snowpark.modin.pandas.window.Window,
@@ -571,9 +567,7 @@ class TelemetryMeta(type):
             attrs (Dict[str, Any]): The attributes of the class.
 
         Returns:
-            Union[snowflake.snowpark.modin.pandas.series.Series,
-                snowflake.snowpark.modin.pandas.dataframe.DataFrame,
-                snowflake.snowpark.modin.pandas.groupby.DataFrameGroupBy,
+            Union[snowflake.snowpark.modin.pandas.groupby.DataFrameGroupBy,
                 snowflake.snowpark.modin.pandas.resample.Resampler,
                 snowflake.snowpark.modin.pandas.window.Window,
                 snowflake.snowpark.modin.pandas.window.Rolling]:
