+++ tests/unit/modin/test_unsupported.py
@@ -67,8 +67,6 @@ def test_unsupported_general(general_method, kwargs):
         ["eval", {"expr": "xxx"}],
         ["ewm", {}],
         ["filter", {}],
-        ["from_dict", {"data": ""}],
-        ["from_records", {"data": ""}],
         ["hist", {}],
         ["interpolate", {}],
         ["isetitem", {"loc": "", "value": ""}],
